Refactor NetworkResult package and add new error string
- Changed the package declaration of NetworkResult.kt from `com.ibrahimkurt.core.network.calladapter` to `com.ibrahimkurt.core.common.network`.
- Added the import of the common `R` resource class in NetworkResult.kt and removed the previous core network `R` resource import.
- Added a new error string "error_unknown" in strings.xml which says "Bilinmeyen bir hata oluştu, bir süre sonra tekrar deneyiniz." (English: "An unknown error occurred, please try again later.").
- Imported `NetworkResult` from the new common network package in NetworkResultCall.kt, NetworkResultCallAdapter.kt, and NetworkResultCallAdapterFactory.kt to ensure consistency across network handling classes.
