+++ core/common/src/main/java/com/ibrahimkurt/core/common/network/NetworkResult.kt
@@ -1,8 +1,8 @@
-package com.ibrahimkurt.core.network.calladapter
+package com.ibrahimkurt.core.common.network
 
 import androidx.annotation.StringRes
+import com.ibrahimkurt.core.common.R
 import com.ibrahimkurt.core.common.util.Constants.EMPTY_STRING
-import com.ibrahimkurt.core.network.R
 
 sealed interface NetworkResult<out T> {
     data class Success<T>(val data: T) : NetworkResult<T>
