+++ core/common/src/main/res/values/strings.xml.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="error_unknown">Bilinmeyen bir hata oluştu, bir süre sonra tekrar deneyiniz.</string>
+</resources>
